requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart
opencv-python-headless
numpy
pillow
openpyxl
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-multipart